requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# MFA & OAuth
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
httpx==0.25.1

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.1

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Production server
gunicorn==21.2.0

# Redis (optional)
redis==5.0.1

# Celery for background tasks (optional)
celery==5.3.4
